requirements.txt
# Multi-Agent Company Intelligence System

langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
streamlit>=1.30.0
python-dotenv>=1.0.0
yfinance>=0.2.36
pandas>=2.0.0
pydantic>=2.0.0
